scraper-api/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
playwright>=1.40
google-api-python-client>=2.0
google-auth>=2.0
firebase-admin>=6.0
anthropic>=0.30
Pillow>=10.0
google-cloud-vision>=3.0
PyMuPDF>=1.23
